requirements.txt
future~=0.18.2
opencv-python~=4.6.0.66
numpy~=1.21.6
Pillow~=9.5.0
torch>=1.13.0
torchvision>=0.14.0
